Boolean.jl Documentation

Overview

This module contains functions to compare Boolean functions. It does this by using a bit-vector representation and comparing bits. The drawback with this representation is that it grows exponentially with the number of variables in a boolean expression.

High level constructors take an algebraic formula as the user representation of the logic function. This is then parsed and converted to a BitVector representing the output of all possible inputs to its "truth table".

Note: When parsing, the operator precedence is respected, with the same precedence that Julia obeys. In particular, xor has the same precedence as or.

There is an associated Jupyter notebook at src/BoolTest.ipynb.

Types

Boolean.Op — Type

Define an operations type. Meant for the operator symbols: :+, :*, :⊕, :~, so that we may dispatch on them as types.

Boolean.Blogic — Type

Structure used to represent a boolean formula involving variables given by a single base string followed by a number.

Note: The formula to be represented must only contain the operators:

~ – The NOT operator.

* – The AND operator.

+ – The OR operator.

⊕ – The XOR operator.

⟹ – The implication operator.

⟺ – The equivalence operator.

along with variables (x1, x2, ...; any base name, one per formula), the constants 0 and 1, and parentheses. Nothing else is accepted.

The first 4 operators are left associative while the last two are right associative. The operator precedence from highest to lowest is:

~

*

+, xor

⟹, ⟺

In practice, one uses a higher level constructor (create_bool_rep) or uses the macro @bfunc. Both of which, in turn, use the inner constructor.

A Blogic is self contained: its truth table is stored in val, whose length is $2^{\rm size}$. Row i of the truth table (entry val[i]) is the value of the formula for the inputs given by the bits of i - 1 (variable 1 is the least significant bit). The number of variables of a formula is the largest variable index it uses, or the minimum set with init_logic, whichever is larger.

Fields

formula :: String – The string representation of the formula.

var :: String – The base name of the logical variables.

size :: Int – The number of variables of the truth table.

val :: BitVector – The bit vector representing the formula. It essentially expresses the values of all

possible inputs.

Constructors

Blogic(form::String, v::String, value::BitVector)

Examples

julia> Blogic("(z1 + z2) * z3", "z", BitVector([0, 0, 0, 0, 0, 1, 1, 1]))
Formula    = (z1 + z2) * z3
Variable   = z
Size       = 3
Bit vector = Bool[0, 0, 0, 0, 0, 1, 1, 1]

This is the logic (boolean) formula that ORs z1 and z2, then ANDs that with z3.

Alternative Blogic Constructor

Boolean.Blogic — Method

Blogic(s[; simplify=false, nvars=0])

Outer constructor for Blogic.

Arguments

s :: String – A logic formula string.

Keyword Arguments

simplify=false::Bool – If true, the logic is simplified before evaluation.

nvars=0::Int – The minimum number of variables of the truth table. The number of variables used is the largest of nvars, the highest variable index in the formula, and the value set with init_logic.

Return

::Blogic

Boolean.Blogic_from_file — Function

Blogic_from_file(f[; simplify=false, nvars=0])

Outer constructor for Blogic.

Arguments

f :: String – A string representing a utf-8 text file containing a logic formula.

Keyword Arguments

simplify=false::Bool – If true, the logic is simplified before evaluation.

nvars=0::Int – The minimum number of variables of the truth table (see Blogic).

Return

::Blogic

Boolean.create_bool_rep — Function

create_bool_rep(s; simplify=false, nvars=0)

Turn boolean formula into a BitVector representation, Blogic.

This is done by the following procedure:

Determine the underlying base variable used in the formula.

Parse the formula into an expression, Expr.

Optionally simplify the logical expression.

Walk the expression tree evaluating it over the BitVector representations of the variables to create the truth table BitVector.

Arguments

s :: String – A logical string.

Keyword Arguments

simplify=false :: Bool – If true simplify the logical expression before creating the BitVector.

nvars=0::Int – The minimum number of variables of the truth table (see Blogic).

Examples

julia> create_bool_rep("(z1 + z2) * z3")
Formula    = (z1 + z2) * z3
Variable   = z
Size       = 3
Bit vector = Bool[0, 0, 0, 0, 0, 1, 1, 1]

Return

::Blogic – Type representing the logical expression.

High Level Module Functions

Boolean.@bfunc — Macro

@bfunc(x)

A macro to create a Blogic function in a syntactically clean way. This macro determines if an input expression is a valid formula and creates the associated "truth table" BitVector based on the number of variables in the formula (see Blogic). Multi-line formulas are entered using a begin/end block. However, each line must be a parse-able expression. So, to connect complicated logic use binary operators on a line by themselves. See the example below.

Examples

julia> @bfunc (z1 + z2) * z3
Formula    = (z1  +  z2)  *  z3
Variable   = z
Size       = 3
Bit vector = Bool[0, 0, 0, 0, 0, 1, 1, 1]

julia> @bfunc begin
       (z1 + z2) * z3
       ⟹
       z4 + z5
       end
Formula    = (z1  +  z2)  *  z3  ⟹  z4  +  z5
Variable   = z
Size       = 5
Bit vector = Bool[1, 1, 1, 1, 1, 0, 0, 0, 1, 1  …  1, 1, 1, 1, 1, 1, 1, 1, 1, 1]

Boolean.Blogic — Method

(Blogic)(xs::Vararg{Integer})

Uses the structure Blogic as a Boolean function.

Arguments

xs :: Vararg{Integer} – A Varargs structure representing inputs (0/1 or Bool) to the Blogic function, f; one per variable of f.

Return

::Bool

Boolean.Blogic — Method

(Blogic)(xm::AbstractMatrix{<:Integer})

Uses the structure Blogic as a Boolean function.

Arguments

xm :: AbstractMatrix{<:Integer} – A matrix of size M, N representing M sets of inputs (0/1 or Bool) to the function, f, which takes N variables.

Return

::BitVector of length M.

Boolean.isEquiv — Method

isEquiv(f1::String, f2::String)

Determines if two logical functions are equivalent when represented as strings.

Arguments

f1 :: String – Formula 1.

f2 :: String – Formula 2.

Return

::Bool – true if the formulas are equivalent; false otherwise.

Boolean.isEquiv — Method

isEquiv(f1::Blogic, f2::Blogic)

Determines if two logical functions are equivalent when represented as Blogic structures. Functions with different numbers of variables are compared over the larger set of variables.

Arguments

f1 :: Blogic – Formula 1.

f2 :: Blogic – Formula 2.

Return

::Bool – true if the formulas are equivalent; false otherwise.

Base Overloaded Functions

Used to show Blogic structures and compare Symbols with Expressions.

Base.show — Method

Base.show(io::IO, x::Blogic)

Show a Blogic structure: the compact form is Blogic("formula"); the text/plain form (used by the REPL) lists the formula, variable, size, and bit vector.

Base.:(==) — Method

Base.:(==)

Overload the equality function for structures of Blogic type.

Low Level Functions

Boolean.logicCount — Function

logicCount(f)

Count the number of true values possible in a given formula.

Arguments

f :: Blogic – A logic formula

Return

The number of true values that are possible with this formula.

Boolean.nonZero — Function

nonZero(f; head=1)

Get up to head inputs that generate true values for a logic function, f.

Arguments

f :: Blogic – A logic formula.

Keyword Arguments

head=1 :: Int – The maximum number of inputs to consider.

Return

::Union{BitMatrix, Nothing} – A matrix whose rows are up to head input values (one column per variable) that will give the logic function, f, a value of true; nothing if there are none.

Boolean.get_non_zero_inputs — Function

get_non_zero_inputs(v, n[; num=1])

Get up to num inputs that generate true values for a logic function. v is a boolean vector that indicates which elements of the truth table yield a value of true.

Arguments

v :: BitVector – A bit vector representing true and false values.

n :: Int – Describes the length of the truth table column: $2^n$.

Keyword Arguments

num :: Int – The desired (maximum) number of inputs that generate truth values.

Return

::Union{BitMatrix, Nothing} – Input values (one row per input, one column per variable) that generate truth values for the function; nothing if there are none.

Boolean.bool_var_rep — Function

bool_var_rep(n)

Generate the boolean bit vectors necessary to represent a logic formula of n variables.

Essentially, generate the truth table of each of the variables collectively as a BitArray.

Arguments

n : Number of logical variables (1 <= n <= 22).

Return

::BitMatrix – The bit representation of all of the logical variables: a matrix of shape (2^n, n), where column j represents variable j.

Boolean.init_logic — Function

init_logic(n)

Sets the minimum number of variables, n, used for the truth tables of formulas built from strings or with @bfunc (a formula always uses at least as many variables as its highest variable index). By default there is no minimum.

Arguments

n :: Int – The minimum number of boolean variables used for formulas (0 <= n <= 22).

Return

Nothing

Boolean.parseLogic — Function

parseLogic

This function creates a parse tree for a boolean expression. The Julia function Meta.parse does this – for the most part. One of the nice features of this parser is that it collapses expressions like (+ (x1 (+ x2 (+x3 x4)))) to (+ x1 x2 x3 x4). It does the same for '*'. However, it does not do so for 'xor'. We adjust this parse tree from Meta.parse so that it does have this property for 'xor'. We also handle the implication operator and logical equivalence operator, by replacing them with their equivalents in terms of ~, +, or *.

The resulting tree is validated: only the operators ~, *, +, ⊕, variables of the form r"[a-zA-Z]+[0-9]+", and the constants 0 and 1 may appear; anything else raises an ArgumentError.

Arguments

expr::String – A logic formula

Return

A parse tree with variable string names replaced with symbols.

Boolean.modifyLogicExpr! — Function

modifyLogicExpr!(e[, n])

Walk an expression tree, converting the logic operators to the (broadcast) Julia operators and variables into their BitVector truth table representations (for n variables; by default the highest variable index in the expression, or the minimum set with init_logic). The result is a Julia expression that evaluates to the truth table of the formula. (The package itself evaluates formulas directly, see evaluate_logic; this function is provided for inspection.)

Arguments

e :: Expr – A (parsed) logic expression.

Return

::Expr – A Julia expression over BitVectors.

Boolean.simplifyLogic — Function

simplifyLogic(e)

Simplify a logical expression.

This function calls a number of specialized variations of this function to deal with different logical operators.

Arguments

e :: Expr – Logic expression.

Return

::Expr – Simplified logic expression (an Int if it reduces to a constant, a Symbol if to a variable).

simplifyLogic(::Op{:+}, xargs::Vector{Any})

simplifLogic for the OR operator.

simplifyLogic(::Op{:*}, xargs::Vector{Any})

simplifyLogic for the AND operator.

simplifyLogic(::Op{:⊕}, xargs::Vector{Any})

simplifyLogic for the XOR operator.

simplifyLogic(e::Union{Int, Symbol})

simplifyLogic for the irreducible cases: A number or a symbol.

Boolean.rle — Function

rle(xs)

Performs a R(un) L(ength) E(ncoding) on an array, grouping like values into arrays.

The values are assumed to be sorted.

Arguments

xs :: Vector{T} – An array that is sortable.

Return

::Vector{Tuple{T, Int}} – A Vector of pairs of the form: (T, Int) representing values from xs and the number of their occurrences.

Boolean.redux — Function

redux(::Op{T}, Tuple{S, Int})

Reduce a pair consisting of an expression and its count to just an expression.

The default case is to just return the expression.

Arguments

::Op{T} – An operator type.

pair :: Tuple{Expr, Int} – Expression and its count.

Return

::Expr – Simplified logic expression.

redux(::Op{:⊕}, pair::Tuple{Expr, Int})

Reduce a pair consisting of an expression and its count to just an expression.

For an XOR expression, we know that only the expression remains or the value is 0.

Arguments

:::Op{:⊕} – An operator type.

pair :: Tuple{Expr, Int} – Expression and its count.

Return

::Expr – Simplified logic expression.

Boolean.evaluate_logic — Function

evaluate_logic(e, n)

Evaluate a (parsed) logic expression tree over the truth tables of n variables, producing the truth table (a BitVector of length $2^n$) of the formula. Constants evaluate to all-false / all-true vectors.

Boolean.validate_single_variable — Function

validate_single_variable(s)

Check that a logic formula string uses only one base variable name. Returns the base variable name string.

Arguments

s :: String – A logic formula string.

Return

::String – The base variable name.

Index

Boolean.Blogic

Boolean.Blogic

Boolean.Blogic

Boolean.Blogic

Boolean.Op

Base.:(==)

Base.show

Boolean.Blogic_from_file

Boolean.bool_var_rep

Boolean.create_bool_rep

Boolean.evaluate_logic

Boolean.get_non_zero_inputs

Boolean.init_logic

Boolean.isEquiv

Boolean.isEquiv

Boolean.logicCount

Boolean.modifyLogicExpr!

Boolean.nonZero

Boolean.parseLogic

Boolean.redux

Boolean.rle

Boolean.simplifyLogic

Boolean.validate_single_variable

Boolean.@bfunc