Boolean.jl Documentation

Overview

This module contains functions to compare Boolean functions. It does this by using a bit-vector representation and comparing bits. The drawback with this representation is that it grows exponentially with the number of variables in a boolean expression.

High level constructors take an algebraic formula as the user representation of the logic function. This is then parsed and converted to a BitVector representing the output of all possible inputs to its "truth table".

Note: When parsing, the operator precedence is respected, with the same precedence that Julia obeys. In particular, xor has the same precedence as or.

There is an associated Jupyter notebook at src/BoolTest.ipynb.

Types

Boolean.Op — Type

Define an operations type. Meant for the operator symbols: :+, :*, :⊕, :~, so that we may dispatch on them as types.

Boolean.Blogic — Type

Structure used to represent a boolean formula involving variables given by a single base string followed by a number.

Note: The formula to be represented must only contain the operators:

~ – The NOT operator.

* – The AND operator.

+ – The OR operator.

⊕ – The XOR operator.

⟹ – The implication operator.

⟺ – The equivalence operator.

The first 4 operators are left associative while the last two are right associative. The operator precedence from highest to lowest is:

~

*

+, xor

⟹, ⟺

In practice, one uses a higher level constructor (create_bool_rep) or uses the macro @bfunc. Both of which, in turn, use the inner constructor.

Fields

formula :: String – The string representation of the formula.

var :: String – The base name of the logical variables.

size :: Int – The number of variables in the formula.

val :: BitVector – The bit vector representing the formula. It essentially expresses the values of all

possible inputs.

Constructors

Blogic(form::String, v::String, value::BitVector)

Examples

julia> Blogic("(z1 + z2) * z3", "z", BitVector([0, 0, 0, 0, 0, 1, 1, 1]))

Formula    = (z1 + z2) * z3
Variable   = z
Size       = 3
Bit vector = Bool[0, 0, 0, 0, 0, 1, 1, 1]

This is the logic (boolean) formula that ORs z1 and z2, then ANDs that with z3.

Alternative Blogic Constructor

Boolean.Blogic — Method

Blogic(s[; simplify])

Outer constructor for Blogic.

Arguments

s :: String – A logic formula string.

Optional Arguments

simplify=false::Bool - Bool argument, if true, logic should be simplified.

Return

::Blogic

Boolean.Blogic_from_file — Method

Blogic_from_file(f[; simplify])

Outer constructor for Blogic.

Arguments

f :: String – A string representing a utf-8 text file containing a logic formula.

Return

::Blogic

Boolean.create_bool_rep — Function

create_bool_rep(s, simplify=false)

Turn boolean formula into a BitVector representation, Blogic.

This is done by the following procedure:

Determine the underlying base variable used in the formula.

Parse the formula into an expression, Expr.

Optionally simplify the logical expression.

Walk the expression tree creating a new tree with Julia mathematical operators substituted for user operators.

Evaluate the expression to create a BitVector.

Arguments

s :: String – A logical string.

simplify=false :: Bool – If true simplify the logical expression before creating the BitVector.

Examples

julia> create_bool_rep("(z1 + z2) * z3")

Formula    = (z1 + z2) * z3
Variable   = z
Size       = 3
Bit vector = Bool[0, 0, 0, 0, 0, 1, 1, 1]

Return

::Blogic – Type representing the logical expression.

High Level Module Functions

Boolean.@bfunc — Macro

@bfunc(x)

A macro to create a Blogic function in a syntactically clean way. This macro determes if an input expression is a valid formula and creates the associated "truth table" BitVectors based on the number of variables in the formula. The function that does this is init_logic which modifies global variables. Multi-line formulas are entered using a begin/end block. However, each line must be a parse-able expression. So, to connect complicated logic use binary operators on a line by themselves. See the example below.

Examples

julia> @bfunc (z1 + z2) * z3

Formula    = (z1 + z2) * z3
Variable   = z
Size       = 3
Bit vector = Bool[0, 0, 0, 0, 0, 1, 1, 1]

julia> @bfunc begin
   (z1 + z2) * z3
   ⟹
   z4 + z5
   end

Formula    = (z1 + z2) * z3 ⟹  z4 + z5
Variable   = z
Size       = 5
Bit vector = Bool[1, 1, 1, 1, 1, 0, 0, 0, 1, 1  …  1, 1, 1, 1, 1, 1, 1, 1, 1, 1]

Boolean.Blogic — Method

(Blogic)(xs::Vararg{Int})

Uses the structure Blogic as a Boolean function.

Arguments

xm :: Vararg{Int} – A Varargs structure representing inputs to the Blogic function, f.

Return

::Bool

Boolean.Blogic — Method

(Blogic)(xm::Matrix{Int})

Uses the structure Blogic as a Boolean function.

Arguments

xm :: Matrix{Int} – A matrix of size M, N representing M sets of inputs to the function, f, which takes N variables.

Return

::BitVector of length M.

Boolean.isEquiv — Method

isEquiv(f1::String, f2::String)

Determines if two logical functions are equivalent when represented as strings.

Arguments

f1 :: String – Formula 1.

f2 :: String – Formula 2.

Return

::Bool – true if the formulas are equivalent; false otherwise.

Boolean.isEquiv — Method

isEquiv(f1::Blogic, f2::Blogic)

Determines if two logical functions are equivalent when represented as Blogic structures.

Arguments

f1 :: Blogic – Formula 1.

f2 :: Blogic – Formula 2.

Return

::Bool – true if the formulas are equivalent; false otherwise.

Base Overloaded Functions

Used to show Blogic structures and compare Symbols with Expressions.

Base.show — Method

Show the Blogic structure.
Params:
io: IO handle.
x : Blogic structure.

Base.show — Method

Show a BitMatrix.

Base.isless — Method

Base.isless(i1::Int, s2::Symbol)

Compare an Int with a Symbol.

Base.isless — Method

Base.isless(i1::Int, e2::Expr)

Compare an Int with an Expr.

Base.isless — Method

Base.isless(s1::Symbol, e2::Expr)

Compare an Symbol with an Expr.

Base.isless — Method

Base.isless(e1::Expr, e2::Expr)

Compare two Julia Expr expressions.

Base.:== — Method

Base.:(==)

Overload the equality function for structures of Blogic type.

Low Level Functions

Boolean.logicCount — Function

logicCount(f)

Count the number of true values possible in a given formula.

Arguments

f :: Blogic – A logic formula

Return

The number of true values that are possible with this formula.

Boolean.nonZero — Function

nonZero(f, head=1)

Get up to head inputs that generate true values for a logic function, f.

Arguments

f :: Blogic – A logic formula.

Optional Arguments

head=1 :: Int – The maximum number of inputs to consider.

Return

A list of up to head input values that will give the logic function, f, a value of true.

Boolean.get_non_zero_inputs — Function

get_non_zero_inputs(v, n[; num=1])

Get num inputs that generate true values for a logic function. v is a boolean vector that indicates which elements of the truth table yield a value of true.

Arguments

v :: BitVector – A bit vector representing true and false values.

n :: Int – Describes the length of the truth table column: $2^n$.

Optional Arguments

num :: Int – The desired number of inputs that generate truth values.

Returns

::Union{BitMatrix, Nothing} – Input values that generate truth values for the current function.

Boolean.bool_var_rep — Function

bool_var_rep(n)

Generate the boolean bit vectors necessary to represent a logic formula of n variables.

Essentially, generate the truth table of each of the variables collectively as a BitArray.

Arguments

n : Number of logical variables.

Return

::BitArray – The bit representation of all of the logical variables.

Boolean.init_logic — Function

init_logic

This sets two global variables, the size of the boolean vectors and the other the Bitarray representations of the variables.

Arguments

n :: Int – The number of boolean variables used in the formulas

this module will consider.

Return

Nothing

Boolean.parseLogic — Function

parseLogic

This function creates a parse tree for a boolean expression. The Julia function Meta.parse does this – for the most part. One of the nice features of this parser is that it collapses expressions like (+ (x1 (+ x2 (+x3 x4)))) to (+ x1 x2 x3 x4). It does the same for '*'. However, it does not do so for 'xor'. We adjust this parse tree from Meta.parse so that it does have this property for 'xor'. We also handle the implication operator and logical equivalence operator, by replacing them with their equivalents in terms of ~, +, or *.

Arguments

expr::String – A logic formula

Return

A parse tree with variable string names replaced with symbols.

Boolean.modifyLogicExpr! — Function

modifyLogicExpr!(e)

The default rule for modifying a logic expression is to do nothing.

modifyLogicExpr!(e::Expr)

Walk an expression tree, converting variable names and operators to Julia operators and variables into BitVector representations.

Arguments

e :: Expr – An expression.

Return

::Expr – A logic expression.

modifyLogicExpr!(e::Symbol)

If e is a Symbol, it should be a variable of the form r"[a-zA-Z]+[0-9]+".

The code splits the name off and uses the number to look up the BitVector representation. Otherwise, it is assumed to be an operator symbol and it is then mapped to the appropriate Julia operator.

NOTE: This will work even if one makes a mistake and uses x3, or y3, the bit vector for the third "variable" will be used.

Arguments

e :: Symbol – An variable or operator.

Return

::Expr – A logic expression.

Boolean.simplifyLogic — Function

simplifyLogic(e)

Simplify a logical expression.

This function calls a number of specialized variations of this function to deal with different logical operators.

Arguments

e :: Expr – Logic expression.

Return

::Expr – Simplified logic expression.

simplifyLogic(::Op{:~}, xargs::Any)

simplifyLogic for the NOT operator.

simplifyLogic(::Op{:+}, xargs::Vector{Any})

simplifLogic for the OR operator.

simplifyLogic(::Op{:*}, xargs::Vector{Any})

simplifyLogic for the AND operator.

simplifyLogic(::Op{:⊕}, xargs::Vector{Any})

simplifyLogic for the XOR operator.

simplifyLogic(e::Union{Int, Symbol})

simplifyLogic for the irreducible cases: A number or a symbol.

Boolean.rle — Function

rle(xs)

Performs a R(un) L(ength) E(ncoding) on an array, grouping like values into arrays.

The values are assumed to be sorted.

Arguments

xs :: Vector{T} – An array that is sortable.

Return

::Vector{Tuple{T, Int}} – A Vector of pairs of the form: (T, Int) representing values from xs and the number of their occurrences.

Boolean.redux — Function

redux(::Op{T}, Tuple{S, Int})

Reduce a pair consisting of an expression and its count to just an expression.

The default case is to just return the expression.

Arguments

::Op{T} – An operator type.

pair :: Tuple{Expr, Int} – Expression and its count.

Return

::Expr – Simplified logic expression.

redux(::Opt{:⊕}, pair::Tuple{Expr, Int})

Reduce a pair consisting of an expression and its count to just an expression.

For an XOR expression, we know that only the expression remains or the value is 0.

Arguments

:::Opt{:⊕} – An operator type.

pair :: Tuple{Expr, Int} – Expression and its count.

Return

::Expr – Simplified logic expression.

Index

Boolean.Blogic

Boolean.Blogic

Boolean.Blogic

Boolean.Blogic

Boolean.Op

Base.:==

Base.isless

Base.isless

Base.isless

Base.isless

Base.show

Base.show

Boolean.Blogic_from_file

Boolean.bool_var_rep

Boolean.create_bool_rep

Boolean.get_non_zero_inputs

Boolean.init_logic

Boolean.isEquiv

Boolean.isEquiv

Boolean.logicCount

Boolean.modifyLogicExpr!

Boolean.nonZero

Boolean.parseLogic

Boolean.redux

Boolean.rle

Boolean.simplifyLogic

Boolean.@bfunc